requirements.txt
streamlit~=1.15.2
pandas~=1.5.2
plotly~=5.11.0
Pillow~=9.3.0